added function getTicketPrice in ShowTimeService class

+++ movies/src/main/java/com/cinemas/movies/entity/ShowTime.java
@@ -9,7 +9,7 @@ public interface ShowTime {
 	Date getShowTime();
 	boolean is3D();
 	long getTheaterScreenCapacity();
-	long getTicketsRemaining();
+	int getTicketsRemaining();
 	int getTicketPrice();
 	String getMovieName();
 	String getTheaterName();
